+++ src/main/java/io/fabric8/maven/docker/util/DockerFileUtil.java
@@ -17,6 +17,7 @@ package io.fabric8.maven.docker.util;/*
 
 import java.io.*;
 import java.util.ArrayList;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.regex.Matcher;
@@ -42,21 +43,22 @@ public class DockerFileUtil {
     private DockerFileUtil() {}
 
     /**
-     * Extract the base image from a dockerfile. The first line containing a <code>FROM</code> is
+     * Extract the base images from a dockerfile. All lines containing a <code>FROM</code> is
      * taken.
      *
      * @param dockerFile file from where to extract the base image
      * @param interpolator interpolator for replacing properties
+     * @return LinkedList of base images name or empty collection if none is found.
      */
-    public static String extractBaseImage(File dockerFile, FixedStringSearchInterpolator interpolator) throws IOException {
+    public static List<String> extractBaseImages(File dockerFile, FixedStringSearchInterpolator interpolator) throws IOException {
         List<String[]> fromLines = extractLines(dockerFile, "FROM", interpolator);
-        if (!fromLines.isEmpty()) {
-            String[] parts = fromLines.get(0);
-            if (parts.length > 1) {
-                return parts[1];
+        LinkedList<String> result = new LinkedList<>();
+        for (String[] fromLine :  fromLines) {
+            if (fromLine.length > 1) {
+                result.add(fromLine[1]);
             }
         }
-        return null;
+        return result;
     }
 
     /**