+++ src/test/java/io/fabric8/maven/docker/service/BuildServiceTest.java
@@ -3,11 +3,15 @@ package io.fabric8.maven.docker.service;
 import java.io.File;
 import java.util.Collections;
 
+import java.util.Properties;
+
+
 import io.fabric8.maven.docker.access.BuildOptions;
 import io.fabric8.maven.docker.access.DockerAccess;
 import io.fabric8.maven.docker.access.DockerAccessException;
 import io.fabric8.maven.docker.assembly.DockerAssemblyManager;
 import io.fabric8.maven.docker.config.BuildImageConfiguration;
+import io.fabric8.maven.docker.util.DockerFileUtilTest;
 import io.fabric8.maven.docker.config.ImageConfiguration;
 import io.fabric8.maven.docker.util.Logger;
 import io.fabric8.maven.docker.util.MojoParameters;
@@ -18,6 +22,7 @@ import mockit.Mocked;
 import mockit.Tested;
 import mockit.Verifications;
 import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.project.MavenProject;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -45,6 +50,15 @@ public class BuildServiceTest {
 
     @Mocked
     private MojoParameters params;
+
+    @Mocked
+    Logger logger;
+
+    @Mocked
+    MojoParameters mojoParameters;
+
+    @Mocked
+    MavenProject mavenProject;
 
     @Injectable
     private QueryService queryService;
@@ -97,6 +111,43 @@ public class BuildServiceTest {
         whenBuildImage(false, false);
         thenImageIsBuilt();
         thenOldImageIsNotRemoved();
+    }
+
+    @Test
+    public void testMultiStageBuild() throws Exception {
+        BuildImageConfiguration buildConfig = new BuildImageConfiguration.Builder()
+                .cleanup("false")
+                .dockerFile(DockerFileUtilTest.class.getResource("Dockerfile_multi_stage").getPath())
+                .filter("false")
+                .build();
+
+        buildConfig.initAndValidate(logger);
+
+        imageConfig = new ImageConfiguration.Builder()
+                .name("build-image")
+                .alias("build-alias")
+                .buildConfig(buildConfig)
+                .build();
+
+        final ImagePullManager pullManager = new ImagePullManager(null,null, null);
+        final BuildService.BuildContext buildContext = new BuildService.BuildContext.Builder()
+                .mojoParameters(mojoParameters)
+                .build();
+
+        new Expectations(mojoParameters) {{
+            mojoParameters.getProject(); result = mavenProject;
+            mavenProject.getProperties(); result = new Properties();
+        }};
+
+        buildService.buildImage(imageConfig, pullManager, buildContext);
+
+        //verify that tries to pull both images
+        new Verifications() {{
+            queryService.hasImage("fabric8/s2i-java");
+            registryService.pullImageWithPolicy("fabric8/s2i-java",  pullManager, buildContext.getRegistryConfig(), false);
+            queryService.hasImage("fabric8/s1i-java");
+            registryService.pullImageWithPolicy("fabric8/s1i-java",  pullManager, buildContext.getRegistryConfig(), false);
+        }};
     }
 
     private void givenAnImageConfiguration(Boolean cleanup) {