+++ doc/changelog.md
@@ -12,7 +12,8 @@
   - Support docker SHELL setting for runCmds (#1157)
   - Added 'autoRemove' option for running containers (#1179)
   - Added support for AWS EC2 instance roles when pushing to AWS ECR (#1186)
-
+  - Add support for auto-pulling multiple base image for multi stage builds (#1057)
+  
 * **0.28.0** (2018-12-13)
   - Update to JMockit 1.43
   - Compiles with Java 11